chore(suite): extract LoggedOutSidebar
